otus-fest-demo/modules/cube_activity/requirements.txt
azure-functions
requests[socks]
pyodbc
#requests==2.19.1

